# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OCTAGON - Local-first operator assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MemoryEntry {
  id: number;
  title: string;
  content: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface MemoryEntryInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  tags?: string[];
}

export interface MemoryEntryUpdate {
  /** @minLength 1 */
  title?: string;
  content?: string;
  tags?: string[];
}

export type DoctrineRuleScope = typeof DoctrineRuleScope[keyof typeof DoctrineRuleScope];


export const DoctrineRuleScope = {
  local: 'local',
  cloud: 'cloud',
  hybrid: 'hybrid',
} as const;

export interface DoctrineRule {
  id: number;
  name: string;
  description: string;
  priority: number;
  scope: DoctrineRuleScope;
  action: string;
  /** @nullable */
  conditions?: string | null;
  active: boolean;
  createdAt: string;
}

export type DoctrineRuleInputScope = typeof DoctrineRuleInputScope[keyof typeof DoctrineRuleInputScope];


export const DoctrineRuleInputScope = {
  local: 'local',
  cloud: 'cloud',
  hybrid: 'hybrid',
} as const;

export interface DoctrineRuleInput {
  /** @minLength 1 */
  name: string;
  description: string;
  priority: number;
  scope: DoctrineRuleInputScope;
  action: string;
  conditions?: string;
  active?: boolean;
}

export type DoctrineRuleUpdateScope = typeof DoctrineRuleUpdateScope[keyof typeof DoctrineRuleUpdateScope];


export const DoctrineRuleUpdateScope = {
  local: 'local',
  cloud: 'cloud',
  hybrid: 'hybrid',
} as const;

export interface DoctrineRuleUpdate {
  name?: string;
  description?: string;
  priority?: number;
  scope?: DoctrineRuleUpdateScope;
  action?: string;
  conditions?: string;
  active?: boolean;
}

export type WorkflowStatus = typeof WorkflowStatus[keyof typeof WorkflowStatus];


export const WorkflowStatus = {
  pending: 'pending',
  active: 'active',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface Workflow {
  id: number;
  name: string;
  description: string;
  status: WorkflowStatus;
  stepCount: number;
  completedSteps: number;
  tags?: string[];
  createdAt: string;
  updatedAt: string;
}

export type WorkflowDetailStatus = typeof WorkflowDetailStatus[keyof typeof WorkflowDetailStatus];


export const WorkflowDetailStatus = {
  pending: 'pending',
  active: 'active',
  completed: 'completed',
  archived: 'archived',
} as const;

export type WorkflowStepStatus = typeof WorkflowStepStatus[keyof typeof WorkflowStepStatus];


export const WorkflowStepStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  skipped: 'skipped',
} as const;

export type WorkflowStepExecutionMode = typeof WorkflowStepExecutionMode[keyof typeof WorkflowStepExecutionMode];


export const WorkflowStepExecutionMode = {
  local: 'local',
  cloud: 'cloud',
  hybrid: 'hybrid',
} as const;

export interface WorkflowStep {
  id: number;
  workflowId: number;
  order: number;
  action: string;
  /** @nullable */
  description?: string | null;
  status: WorkflowStepStatus;
  /** @nullable */
  result?: string | null;
  executionMode?: WorkflowStepExecutionMode;
}

export interface WorkflowDetail {
  id: number;
  name: string;
  description: string;
  status: WorkflowDetailStatus;
  steps: WorkflowStep[];
  tags?: string[];
  createdAt: string;
  updatedAt: string;
}

export interface WorkflowInput {
  /** @minLength 1 */
  name: string;
  description: string;
  tags?: string[];
}

export type WorkflowUpdateStatus = typeof WorkflowUpdateStatus[keyof typeof WorkflowUpdateStatus];


export const WorkflowUpdateStatus = {
  pending: 'pending',
  active: 'active',
  completed: 'completed',
  archived: 'archived',
} as const;

export interface WorkflowUpdate {
  name?: string;
  description?: string;
  status?: WorkflowUpdateStatus;
  tags?: string[];
}

export type WorkflowStepInputExecutionMode = typeof WorkflowStepInputExecutionMode[keyof typeof WorkflowStepInputExecutionMode];


export const WorkflowStepInputExecutionMode = {
  local: 'local',
  cloud: 'cloud',
  hybrid: 'hybrid',
} as const;

export interface WorkflowStepInput {
  /** @minLength 1 */
  action: string;
  description?: string;
  order: number;
  executionMode?: WorkflowStepInputExecutionMode;
}

export type WorkflowStepUpdateStatus = typeof WorkflowStepUpdateStatus[keyof typeof WorkflowStepUpdateStatus];


export const WorkflowStepUpdateStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  skipped: 'skipped',
} as const;

export type WorkflowStepUpdateExecutionMode = typeof WorkflowStepUpdateExecutionMode[keyof typeof WorkflowStepUpdateExecutionMode];


export const WorkflowStepUpdateExecutionMode = {
  local: 'local',
  cloud: 'cloud',
  hybrid: 'hybrid',
} as const;

export interface WorkflowStepUpdate {
  status?: WorkflowStepUpdateStatus;
  result?: string;
  executionMode?: WorkflowStepUpdateExecutionMode;
}

export type RoutingDecisionDecision = typeof RoutingDecisionDecision[keyof typeof RoutingDecisionDecision];


export const RoutingDecisionDecision = {
  local: 'local',
  cloud: 'cloud',
  hybrid: 'hybrid',
} as const;

export interface RoutingDecision {
  id: number;
  query: string;
  decision: RoutingDecisionDecision;
  localScore: number;
  cloudScore: number;
  reasoning: string;
  /** @nullable */
  latencyMs?: number | null;
  createdAt: string;
}

export interface RoutingQuery {
  /** @minLength 1 */
  query: string;
  context?: string;
}

export type ScenarioStatus = typeof ScenarioStatus[keyof typeof ScenarioStatus];


export const ScenarioStatus = {
  draft: 'draft',
  ready: 'ready',
  running: 'running',
  completed: 'completed',
} as const;

export interface Scenario {
  id: number;
  name: string;
  description: string;
  status: ScenarioStatus;
  steps: string[];
  tags: string[];
  createdAt: string;
}

export interface ScenarioInput {
  /** @minLength 1 */
  name: string;
  description: string;
  steps?: string[];
  tags?: string[];
}

export type ScenarioUpdateStatus = typeof ScenarioUpdateStatus[keyof typeof ScenarioUpdateStatus];


export const ScenarioUpdateStatus = {
  draft: 'draft',
  ready: 'ready',
  running: 'running',
  completed: 'completed',
} as const;

export interface ScenarioUpdate {
  name?: string;
  description?: string;
  status?: ScenarioUpdateStatus;
  steps?: string[];
  tags?: string[];
}

export interface DashboardSummary {
  totalMemoryEntries: number;
  totalDoctrineRules: number;
  totalWorkflows: number;
  activeWorkflows: number;
  completedWorkflows: number;
  totalRoutingDecisions: number;
  localDecisions: number;
  cloudDecisions: number;
  hybridDecisions: number;
  totalScenarios: number;
  readyScenarios: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  memory: 'memory',
  doctrine: 'doctrine',
  workflow: 'workflow',
  routing: 'routing',
  scenario: 'scenario',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  description: string;
  createdAt: string;
}

export interface RoutingStats {
  local: number;
  cloud: number;
  hybrid: number;
  total: number;
  localPct?: number;
  cloudPct?: number;
  hybridPct?: number;
}

export type ListMemoryEntriesParams = {
tag?: string;
search?: string;
};

export type ListWorkflowsParams = {
status?: ListWorkflowsStatus;
};

export type ListWorkflowsStatus = typeof ListWorkflowsStatus[keyof typeof ListWorkflowsStatus];


export const ListWorkflowsStatus = {
  pending: 'pending',
  active: 'active',
  completed: 'completed',
  archived: 'archived',
} as const;