# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OCTAGON - Local-first operator assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memory:
    get:
      operationId: listMemoryEntries
      tags: [memory]
      summary: List memory entries
      parameters:
        - in: query
          name: tag
          schema:
            type: string
          required: false
        - in: query
          name: search
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of memory entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryEntry"
    post:
      operationId: createMemoryEntry
      tags: [memory]
      summary: Create a memory entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryEntryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEntry"
  /memory/{id}:
    get:
      operationId: getMemoryEntry
      tags: [memory]
      summary: Get a memory entry
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Memory entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEntry"
        "404":
          description: Not found
    patch:
      operationId: updateMemoryEntry
      tags: [memory]
      summary: Update a memory entry
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEntry"
    delete:
      operationId: deleteMemoryEntry
      tags: [memory]
      summary: Delete a memory entry
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "204":
          description: Deleted
  /doctrine:
    get:
      operationId: listDoctrineRules
      tags: [doctrine]
      summary: List doctrine rules
      responses:
        "200":
          description: List of doctrine rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DoctrineRule"
    post:
      operationId: createDoctrineRule
      tags: [doctrine]
      summary: Create a doctrine rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctrineRuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctrineRule"
  /doctrine/{id}:
    patch:
      operationId: updateDoctrineRule
      tags: [doctrine]
      summary: Update a doctrine rule
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctrineRuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctrineRule"
    delete:
      operationId: deleteDoctrineRule
      tags: [doctrine]
      summary: Delete a doctrine rule
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "204":
          description: Deleted
...
  /scenarios:
    get:
      operationId: listScenarios
      tags: [scenarios]
      summary: List demo scenarios
      responses:
        "200":
          description: List of scenarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scenario"
    post:
      operationId: createScenario
      tags: [scenarios]
      summary: Create a demo scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
  /scenarios/{id}:
    get:
      operationId: getScenario
      tags: [scenarios]
      summary: Get a scenario
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: Scenario
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
        "404":
          description: Not found
    patch:
      operationId: updateScenario
      tags: [scenarios]
      summary: Update a scenario
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scenario"
...
components:
  schemas:
...
    MemoryEntry:
      type: object
      required: [id, title, content, tags, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MemoryEntryInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
          minLength: 1
        tags:
          type: array
          items:
            type: string
    MemoryEntryUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        tags:
          type: array
          items:
            type: string
    DoctrineRule:
      type: object
      required: [id, name, description, priority, scope, action, active, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
        scope:
          type: string
          enum: [local, cloud, hybrid]
        action:
          type: string
        conditions:
          type: ["string", "null"]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    DoctrineRuleInput:
      type: object
      required: [name, description, priority, scope, action]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        priority:
          type: integer
        scope:
          type: string
          enum: [local, cloud, hybrid]
        action:
          type: string
        conditions:
          type: string
        active:
          type: boolean
    DoctrineRuleUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        priority:
          type: integer
        scope:
          type: string
          enum: [local, cloud, hybrid]
        action:
          type: string
        conditions:
          type: string
        active:
          type: boolean
...
    Scenario:
      type: object
      required: [id, name, description, status, steps, tags, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, ready, running, completed]
        steps:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    ScenarioInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        steps:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
    ScenarioUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft, ready, running, completed]
        steps:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
...